Silently ignore charset escapes
Rust tools such as Cargo and `rustfmt` seem to like resetting the
charset to US ASCII by using `ESC(B` in the middle of a line. Currently
this outputs a literal `(B` in the rendered HTML which makes the output
look broken.

This simply discards the next character after any `ESC(` or `ESC)`.
Initially I only discarded `B` for US ASCII and let the other charsets
through as their original escape sequence. However, the existing pattern
in the code is to ignore any well-formed escape sequences even if their
exact content isn't understood.