Replace bindata with embedded style sheet
I've preserved the nature of the file being gzipped and committed the
gzipped file to the repo. It can be created from its source with `go
generate ./...`, but the standard practice is to commit files generated
by `go generate` so that a `go install` command is able to use them.