Disable cgo on build
Whenever I build this on my machine, it does not work when we use this
in prod. This is because my machine has a newer version of glibc than
the one used in prod.

Disabling CGO should fix this.

To check that this does indeed disable CGO, you can run:
```
go tool nm <path-to-binary> | grep Cfunc
```
on the binary that matches the OS and ARCH that the build was run on.