Eliminate strings.Replace; update tests
Replace was used to put `&nbsp;` inside otherwise blank lines (`\n\n` -> `\n&nbsp;\n`). I went looking through history to understand. It used to be implemented as a regexp replace from `^$`. That originated in the Ruby version going back to ef20b4a, before that there was something funky going on with divs.

Changing it to generate `&nbsp;` on the fly broke two tests. If the rationale for `&nbsp;` is to ensure empty lines aren't lost, then the two tests were arguably wrong:
* "clears everything after the \x1b[0K" clears the "hello" line, but because that blank line is at the start of the output it wasn't between two newlines, missing the Replace
* a similar problem occurs with Pikachu, but at both the start and the end (if you are unconvinced about the end, note there are two blank lines after Pikachu if you cat fixtures/pikachu.sh.raw).